+++ src/main/resources/logback.xml
@@ -19,14 +19,11 @@
   <shutdownHook/>
   <appender name="STDERR" class="ch.qos.logback.core.ConsoleAppender">
     <target>System.err</target>
-    <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
-      <level>warn</level>
-    </filter>
     <encoder>
       <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
     </encoder>
   </appender>
-  <root level="warn">
+  <root level="debug">
     <appender-ref ref="STDERR"/>
   </root>
 </configuration>
